# frontend/openapi/openapi.json
{
  "openapi": "3.1.2",
  "info": {
    "title": "PLIEGO API",
    "description": "API de PLIEGO v1: autenticación, clientes, catálogo, inventario, carrito, pago y pedidos.",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://127.0.0.1:18445",
      "description": "Generated server url"
    }
  ],
...
  "paths": {
...
    "/api/v1/admin/editions/{editionId}/offer": {
      "put": {
        "tags": [
          "Administración de catálogo"
        ],
        "summary": "Establecer una oferta de edición",
        "description": "Reemplaza la única oferta de la edición. PostgreSQL valida el precio inferior al precio base y el intervalo de fechas.",
        "operationId": "set",
        "parameters": [
          {
            "name": "editionId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64",
              "exclusiveMinimum": 0
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EditionOfferRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "204": {
            "description": "Oferta guardada"
          },
          "400": {
            "description": "Precio o fechas inválidos",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            }
          },
          "404": {
            "description": "Edición no encontrada",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerJwt": []
          }
        ]
      },
      "delete": {
        "tags": [
          "Administración de catálogo"
        ],
        "summary": "Desactivar una oferta de edición",
        "description": "Desactiva la oferta sin cambiar el precio base ni los importes históricos de pedidos.",
        "operationId": "clear",
        "parameters": [
          {
            "name": "editionId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64",
              "exclusiveMinimum": 0
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Oferta desactivada"
          },
          "404": {
            "description": "Edición no encontrada",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerJwt": []
          }
        ]
      }
    },
...
    "/api/v1/auth/register": {
      "post": {
        "tags": [
          "Authentication"
        ],
        "summary": "Registrar una cuenta CUSTOMER",
        "description": "Aplica BCrypt y crea cuenta pendiente de verificación junto con el correo en outbox, en una única transacción. Confirma el enlace antes de iniciar sesión.",
        "operationId": "register",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RegisterRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Cuenta de cliente creada",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RegisterResponse"
                }
              }
            }
          },
          "400": {
            "description": "Datos inválidos",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            }
          },
          "409": {
            "description": "Correo ya registrado",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            }
          },
          "500": {
            "description": "Error interno",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/auth/refresh": {
      "post": {
        "tags": [
          "Authentication"
        ],
        "summary": "Restaurar o renovar la sesión",
        "description": "Rota la credencial persistente HttpOnly y emite un access token nuevo. Si se pierde la respuesta, repetir con la cookie anterior durante 5 minutos devuelve la misma rotación. La sesión tiene una expiración absoluta de 30 días desde el inicio de sesión.",
        "operationId": "refreshSession",
        "parameters": [
          {
            "name": "X-PLIEGO-SESSION-REQUEST",
            "in": "header",
            "description": "Marcador requerido para autorizar la petición de sesión desde el cliente PLIEGO.",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Sesión restaurada",
            "headers": {
              "Set-Cookie": {
                "description": "Cookie HttpOnly renovada; conserva la expiración absoluta de la sesión.",
                "style": "simple",
                "schema": {
                  "type": "string"
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LoginResponse"
                }
              }
            }
          },
          "204": {
            "description": "No existe una sesión válida; se elimina la cookie",
            "headers": {
              "Set-Cookie": {
                "description": "Cookie eliminada.",
                "style": "simple",
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "500": {
            "description": "Error interno",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            }
          }
        },
        "security": [
          {
            "authSessionCookie": []
          }
        ]
      }
    },
    "/api/v1/auth/logout": {
      "post": {
        "tags": [
          "Authentication"
        ],
        "summary": "Cerrar sesión",
        "description": "Revoca la sesión persistente y elimina su cookie HttpOnly. La operación es idempotente.",
        "operationId": "logoutSession",
        "parameters": [
          {
            "name": "X-PLIEGO-SESSION-REQUEST",
            "in": "header",
            "description": "Marcador requerido para autorizar la petición de sesión desde el cliente PLIEGO.",
            "required": true
          }
        ],
        "responses": {
          "204": {
            "description": "Sesión revocada y cookie eliminada",
            "headers": {
              "Set-Cookie": {
                "description": "Cookie eliminada.",
                "style": "simple",
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "500": {
            "description": "Error interno",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            }
          }
        },
        "security": [
          {
            "authSessionCookie": []
          }
        ]
      }
    },
    "/api/v1/auth/login": {
      "post": {
        "tags": [
          "Authentication"
        ],
        "summary": "Iniciar sesión",
        "description": "Emite un access token bearer HS256 válido durante 30 minutos y establece una cookie HttpOnly de sesión revocable, válida durante 30 días.",
        "operationId": "login",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Token de acceso emitido y sesión persistente establecida",
            "headers": {
              "Set-Cookie": {
                "description": "Cookie HttpOnly, Secure y SameSite=Strict; expira 30 días después del inicio de sesión.",
                "style": "simple",
                "schema": {
                  "type": "string"
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LoginResponse"
                }
              }
            }
          },
          "400": {
            "description": "Datos inválidos",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            }
          },
          "401": {
            "description": "Credenciales inválidas",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            }
          },
          "403": {
            "description": "Correo pendiente de verificación (EMAIL_NOT_VERIFIED)",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            }
          },
          "500": {
            "description": "Error interno",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ProblemDetail": {
        "type": "object",
        "description": "Detalle de problema RFC 9457 con campos de correlación de PLIEGO.",
        "properties": {
          "type": {
            "type": "string",
            "format": "uri",
            "example": "urn:pliego:problem:VALIDATION_ERROR"
          },
          "title": {
            "type": "string",
            "example": "Datos inválidos"
          },
          "status": {
            "type": "integer",
            "format": "int32",
            "example": 400
          },
          "detail": {
            "type": "string",
            "example": "Revisa los datos enviados e intenta nuevamente."
          },
          "instance": {
            "type": "string",
            "format": "uri",
            "example": "/api/v1/auth/register"
          },
          "code": {
            "type": "string",
            "description": "Código estable legible por máquina. Los errores de dominio usan el SQLSTATE canónico de PostgreSQL (Pxxxx); los errores propios de la capa REST usan un código simbólico (por ejemplo VALIDATION_ERROR o INVALID_CARD_NUMBER).",
            "example": "P3002"
          },
          "traceId": {
            "type": "string",
            "example": "8f2ab598-3e33-40b1-a5c7-15f676a84f9a"
          }
        }
      },
...
      "EditionOfferRequest": {
        "type": "object",
        "properties": {
          "offerPrice": {
            "type": "string",
            "example": "15.00",
            "pattern": "(?:0\\.(?:0[1-9]|[1-9][0-9])|[1-9][0-9]{0,8}\\.[0-9]{2})"
          },
          "startsAt": {
            "type": "string",
            "format": "date-time"
          },
          "endsAt": {
            "type": "string",
            "format": "date-time"
          },
          "offerCopy": {
            "type": "string",
            "maxLength": 1000,
            "minLength": 0
          },
          "terms": {
            "type": "string",
            "maxLength": 5000,
            "minLength": 0
          }
        },
        "required": [
          "endsAt",
          "offerPrice",
          "startsAt"
        ]
      },
...
      "RegisterRequest": {
        "type": "object",
        "description": "Crea una cuenta CUSTOMER pendiente de verificación de correo.",
        "properties": {
          "email": {
            "type": "string",
            "format": "email",
            "example": "usuario@example.com",
            "maxLength": 254,
            "minLength": 0
          },
          "password": {
            "type": "string",
            "format": "password",
            "description": "Al menos 8 caracteres Unicode y máximo 72 bytes en UTF-8.",
            "example": "password-segura",
            "maxLength": 72,
            "minLength": 8
          },
          "firstNames": {
            "type": "string",
            "example": "Ana María",
            "maxLength": 120,
            "minLength": 1
          },
          "lastNames": {
            "type": "string",
            "example": "Pérez López",
            "maxLength": 120,
            "minLength": 1
          },
          "phone": {
            "type": [
              "string",
              "null"
            ],
            "description": "Prefijo internacional obligatorio; admite separadores de presentación.",
            "example": "+59325550134"
          }
        },
        "required": [
          "email",
          "firstNames",
          "lastNames",
          "password"
        ]
      },
      "RegisterResponse": {
        "type": "object",
        "properties": {
          "userId": {
            "type": "string",
            "example": "100",
            "pattern": "^[1-9][0-9]*$"
          },
          "customerId": {
            "type": "string",
            "example": "87",
            "pattern": "^[1-9][0-9]*$"
          },
          "state": {
            "type": "string",
            "description": "Requiere verificar correo antes de iniciar sesión.",
            "enum": [
              "PENDING_VERIFICATION"
            ],
            "example": "PENDING_VERIFICATION"
          }
        }
      },
      "AuthenticatedUser": {
        "type": "object",
        "properties": {
          "userId": {
            "type": "string",
            "example": "100",
            "pattern": "^[1-9][0-9]*$"
          },
          "email": {
            "type": "string",
            "example": "usuario@example.com"
          },
          "role": {
            "type": "string",
            "enum": [
              "CUSTOMER",
              "ADMIN"
            ],
            "example": "CUSTOMER"
          }
        }
      },
      "LoginResponse": {
        "type": "object",
        "properties": {
          "accessToken": {
            "type": "string",
            "format": "jwt"
          },
          "tokenType": {
            "type": "string",
            "example": "Bearer"
          },
          "expiresInSeconds": {
            "type": "integer",
            "format": "int32",
            "example": 1800
          },
          "user": {
            "$ref": "#/components/schemas/AuthenticatedUser"
          }
        }
      },
      "LoginRequest": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "format": "email",
            "example": "usuario@example.com",
            "maxLength": 254,
            "minLength": 0
          },
          "password": {
            "type": "string",
            "format": "password",
            "description": "Máximo 72 bytes en UTF-8.",
            "maxLength": 72,
            "minLength": 1
          }
        },
        "required": [
          "email",
          "password"
        ]
      },
...
    },
    "securitySchemes": {
      "bearerJwt": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      },
      "authSessionCookie": {
        "type": "apiKey",
        "name": "pliego_session",
        "in": "cookie"
      }
    }
  }
}